feat: refactor code for login
