+++ movie-frontend/src/App.js
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 import { Route, Routes } from 'react-router-dom';
 import './App.css';
 import api from './api/axiosConfig';
-import Layout from './components/Layjout';
+import Layout from './components/Layout';
 import Header from './components/header/Header';
 import Home from './components/home/Home';
 import Login from './components/login/Login';
